Lots of K-Type LED cleanup
- Added pixelSCTest cli
- Added pixelXYTest cli
- Initial work on PixelAddressType_Rect
- Alignment fixes for interpolation
- General cleanup of LED test code
- Tuned i2c bus for 900 kBaud
- Lots of LED mapping fixes
- Added more animation tests
- Reduced global brightness

+++ Macro/PixelMap/pixel.h
@@ -178,9 +178,9 @@ extern FrameState Pixel_FrameState;
 
 extern       PixelBuf     Pixel_Buffers[];
 extern const PixelElement Pixel_Mapping[];
-extern const uint8_t      Pixel_DisplayMapping[];
+extern const uint16_t     Pixel_DisplayMapping[];
 extern const uint8_t    **Pixel_Animations[];
-extern const uint8_t      Pixel_ScanCodeToPixel[];
+extern const uint16_t     Pixel_ScanCodeToPixel[];
